Move the registration periods cache into its own module
